+++ packages/docs/blog/2024-05-16-effects.tsx
@@ -1,0 +1,12 @@
+import {makeScene2D} from '@motion-canvas/2d';
+import {createEffect, createSignal} from '@motion-canvas/core';
+
+export default makeScene2D(function* () {
+  const signal = createSignal(0);
+
+  createEffect(() => {
+    console.log('Signal changed: ', signal());
+  });
+
+  yield* signal(1, 2);
+});